[refactor] Extract validateTypeFilter, rewire execValidateFlags and listValidateType (#147)
- Add cmd/flags.go with validateTypeFilter(typeFilter string) error
- listValidateType becomes a one-line shim delegating to validateTypeFilter
- execValidateFlags delegates inner type check to validateTypeFilter
- Eliminates duplicate ValidPrefixType check and error-format string
